+++ test/unit/fuzz-tools.test.js
@@ -226,6 +226,7 @@ import * as estimatorsV233 from '../../lib/estimators-v233.js';
 import * as dermscoreV234 from '../../lib/dermscore-v234.js';
 import * as painscoreV235 from '../../lib/painscore-v235.js';
 import * as ophthoV236 from '../../lib/ophtho-v236.js';
+import * as cardioechoV237 from '../../lib/cardioecho-v237.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -396,6 +397,7 @@ const MODULES = {
   'dermscore-v234.js': dermscoreV234,
   'painscore-v235.js': painscoreV235,
   'ophtho-v236.js': ophthoV236,
+  'cardioecho-v237.js': cardioechoV237,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];